fix(build): move TEAM_SYNC_OBSERVED_TABLES to dependency-free schema-ddl
R4 imported TEAM_SYNC_OBSERVED_TABLES into the migration chain from
db/queries/team-outbox.ts, which imports db/client (bun:sqlite). The
@goondocks/myco-team CLI imports db/schema.js (for SCHEMA_VERSION) → migrations
→ team-outbox → client, so the worker-CLI esbuild bundle could no longer
resolve bun:sqlite and `make dev-build` failed. Move the canonical synced-table
list to the dependency-free schema-ddl module (its natural home, alongside the
delete-trigger list); team-outbox re-exports it for existing consumers and the
migration chain imports it from schema-ddl — breaking the heavy edge. Value is
identical; parity test still green.
